♻️ refactor: profile 컴포넌트 import 상대 경로 였던 부분들 모두 절대 경로로 변경
