+++ client/src/components/profile/header/index.ts
@@ -1,4 +1,4 @@
-export * from "./ui/Avatar";
-export * from "./ui/Banner";
-export * from "./ui/Info";
-export * from "./ui/Stats";
+export * from "@/components/profile/header/ui/Avatar";
+export * from "@/components/profile/header/ui/Banner";
+export * from "@/components/profile/header/ui/Info";
+export * from "@/components/profile/header/ui/Stats";
